Fix playlist not progressing past first song in scenes where time is paused
When entering scenes such as R6D or Astronaut Complex, onGamePause() is called
(probably because the game time is paused). When this happens, the music player
was not properly unpaused again (only SoundtrackEditor.CurrentSituation.paused
was reset, but SoundtrackEditor.PlaybackPaused which is the variable that
is checked to determine if the next song should be started when the current
song ends was not reset).